requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==0.29.0
pandas==2.2.2
openpyxl==3.1.5
faker==19.13.0
networkx==3.2.1
